+++ packages/component-library/src/components/Hero/Hero.stories.tsx
@@ -26,6 +26,7 @@ export default {
     subtitle: { name: 'Subtitle' },
     body: { name: 'Body' },
     image: { name: 'Image' },
+    background: { name: 'Background' },
     actions: { name: 'Actions' }
   }
 };